# perf(mutation): refactor parent removed detection to iterative procedure