--- a/.changeset/kind-kids-design.md
+++ b/.changeset/kind-kids-design.md
@@ -1,0 +1,5 @@
+---
+"rrweb": patch
+---
+
+Optimize performance of isParentRemoved by converting it to an iterative procedure