--- a/.changeset/silent-plants-perform.md
+++ b/.changeset/silent-plants-perform.md
@@ -1,5 +1,5 @@
 ---
-'rrweb': patch
+"rrweb": patch
 ---
 
 Return early for child same origin frames