@Juice10 2024-05-31
Good catch Jonas, could you add a changeset, then I'll merge it