@JonasBa 2024-05-31
I think there is something up with gh ci, that check if failing on a network timeout